Almost done with server functionality
The changes should be transparent to old and new users now.
Except for the fact that we don't check for the cmake capabilities yet.
