+++ commands/__init__.py
@@ -1,21 +1,14 @@
+from .build import CmakeBuildCommand, CmakeExecCommand
 from .clear_cache import CmakeClearCacheCommand
 from .configure import CmakeConfigureCommand
+from .configure2 import CmakeConfigure2Command
 from .diagnose import CmakeDiagnoseCommand
 from .edit_cache import CmakeEditCacheCommand
 from .insert_diagnosis import CmakeInsertDiagnosisCommand
 from .new_project import CmakeNewProjectCommand
 from .open_build_folder import CmakeOpenBuildFolderCommand
+from .reveal_include_directories import CmakeRevealIncludeDirectories
 from .run_ctest import CmakeRunCtestCommand
+from .set_target import CmakeSetTargetCommand
 from .write_build_targets import CmakeWriteBuildTargetsCommand
-
-__all__ = [
-    'CmakeClearCacheCommand'
-    , 'CmakeConfigureCommand'
-    , 'CmakeDiagnoseCommand'
-    , 'CmakeEditCacheCommand'
-    , 'CmakeInsertDiagnosisCommand'
-    , 'CmakeNewProjectCommand'
-    , 'CmakeOpenBuildFolderCommand'
-    , 'CmakeRunCtestCommand'
-    , 'CmakeWriteBuildTargetsCommand'
-]
+from .command import ServerManager
