+++ commands/configure2.py
@@ -1,0 +1,7 @@
+from .command import CmakeCommand
+
+
+class CmakeConfigure2Command(CmakeCommand):
+
+    def run(self):
+        self.server.configure(self.cmake.command_line_overrides)
