+++ event_listeners/__init__.py
@@ -1,3 +1,1 @@
 from .configure_on_save import ConfigureOnSave
-
-__all__ = ['ConfigureOnSave']
